add new base model and modify all model adapter
