+++ ts_benchmark/baselines/pathformer/pathformer.py
@@ -1,26 +1,10 @@
-from typing import Optional, Tuple
-
-import math
-import numpy as np
-import pandas as pd
-import torch
-import torch.nn as nn
-from torch.utils.data import DataLoader
-from sklearn.preprocessing import StandardScaler
-from torch import optim
 from torch.optim import lr_scheduler
 
-from ts_benchmark.baselines.pathformer.models.pathformer_model import PathformerModel
-from ts_benchmark.baselines.utils import (
-    forecasting_data_provider,
-    train_val_split,
-    get_time_mark,
-)
-from ts_benchmark.utils.data_processing import split_time
+from ts_benchmark.models.advanced_model_base import Advanced_Model_Base
+from .models.pathformer_model import PathformerModel
 from .utils.tools import EarlyStopping, adjust_learning_rate
-from ...models.model_base import ModelBase, BatchMaker
 
-DEFAULT_HYPER_PARAMS = {
+MODEL_HYPER_PARAMS = {
     "k": 2,
     "enc_in": 1,
     "dec_in": 1,
@@ -69,576 +53,39 @@ DEFAULT_HYPER_PARAMS = {
     "batch_norm": 0,
 }
 
-
-class PathformerConfig:
+class Pathformer(Advanced_Model_Base):
     def __init__(self, **kwargs):
-        for key, value in DEFAULT_HYPER_PARAMS.items():
-            setattr(self, key, value)
-
-        for key, value in kwargs.items():
-            setattr(self, key, value)
-
-    @property
-    def pred_len(self):
-        return self.horizon
-
-
-class Pathformer(ModelBase):
-    def __init__(self, **kwargs):
-        super(Pathformer, self).__init__()
-        self.config = PathformerConfig(**kwargs)
-        self.scaler = StandardScaler()
-        self.seq_len = self.config.seq_len
-        self.win_size = self.config.seq_len
-
-    @staticmethod
-    def required_hyper_params() -> dict:
-        """
-        Return the hyperparameters required by model.
-
-        :return: An empty dictionary indicating that model does not require additional hyperparameters.
-        """
-        return {
-            "seq_len": "input_chunk_length",
-            "horizon": "output_chunk_length",
-            "norm": "norm",
-        }
+        super(Pathformer, self).__init__(MODEL_HYPER_PARAMS, **kwargs)
+        self.config.adj_lr_in_batch = True if self.config.lradj == "TST" else False
+        self.config.adj_lr_in_epoch = True if self.config.lradj != "TST" else False
 
     @property
     def model_name(self):
         return "Pathformer"
 
-    def multi_forecasting_hyper_param_tune(self, train_data: pd.DataFrame):
-        freq = pd.infer_freq(train_data.index)
-        if freq == None:
-            raise ValueError("Irregular time intervals")
-        elif freq[0].lower() not in ["m", "w", "b", "d", "h", "t", "s"]:
-            self.config.freq = "s"
-        else:
-            self.config.freq = freq[0].lower()
-
-        column_num = train_data.shape[1]
-        self.config.enc_in = column_num
-        self.config.dec_in = column_num
-        self.config.c_out = column_num
-
-        if self.model_name == "MICN":
-            setattr(self.config, "label_len", self.config.seq_len)
-        else:
-            setattr(self.config, "label_len", self.config.seq_len // 2)
-
-    def single_forecasting_hyper_param_tune(self, train_data: pd.DataFrame):
-        freq = pd.infer_freq(train_data.index)
-        if freq == None:
-            raise ValueError("Irregular time intervals")
-        elif freq[0].lower() not in ["m", "w", "b", "d", "h", "t", "s"]:
-            self.config.freq = "s"
-        else:
-            self.config.freq = freq[0].lower()
-
-        column_num = train_data.shape[1]
-        self.config.enc_in = column_num
-        self.config.dec_in = column_num
-        self.config.c_out = column_num
-
-        setattr(self.config, "label_len", self.config.horizon)
-
-    def detect_hyper_param_tune(self, train_data: pd.DataFrame):
-        freq = pd.infer_freq(train_data.index)
-        if freq == None:
-            raise ValueError("Irregular time intervals")
-        elif freq[0].lower() not in ["m", "w", "b", "d", "h", "t", "s"]:
-            self.config.freq = "s"
-        else:
-            self.config.freq = freq[0].lower()
-
-        column_num = train_data.shape[1]
-        self.config.enc_in = column_num
-        self.config.dec_in = column_num
-        self.config.c_out = column_num
-        self.config.label_len = 48
-
-    def padding_data_for_forecast(self, test):
-        time_column_data = test.index
-        data_colums = test.columns
-        start = time_column_data[-1]
-        date = pd.date_range(
-            start=start, periods=self.config.horizon + 1, freq=self.config.freq.upper()
-        )
-        df = pd.DataFrame(columns=data_colums)
-
-        df.iloc[: self.config.horizon + 1, :] = 0
-
-        df["date"] = date
-        df = df.set_index("date")
-        new_df = df.iloc[1:]
-        test = pd.concat([test, new_df])
-        return test
-
-    def _padding_time_stamp_mark(
-        self, time_stamps_list: np.ndarray, padding_len: int
-    ) -> np.ndarray:
-        """
-        Padding time stamp mark for prediction.
-
-        :param time_stamps_list: A batch of time stamps.
-        :param padding_len: The len of time stamp need to be padded.
-        :return: The padded time stamp mark.
-        """
-        padding_time_stamp = []
-        for time_stamps in time_stamps_list:
-            start = time_stamps[-1]
-            expand_time_stamp = pd.date_range(
-                start=start,
-                periods=padding_len + 1,
-                freq=self.config.freq.upper(),
+    def _adjust_lr(self, optimizer, epoch, config):
+        if not hasattr(self, "scheduler"):
+            train_steps = len(self.train_data_loader)
+            self.scheduler = lr_scheduler.OneCycleLR(
+                optimizer=optimizer,
+                steps_per_epoch=train_steps,
+                pct_start=config.pct_start,
+                epochs=config.train_epochs,
+                max_lr=config.learning_rate,
             )
-            padding_time_stamp.append(expand_time_stamp.to_numpy()[-padding_len:])
-        padding_time_stamp = np.stack(padding_time_stamp)
-        whole_time_stamp = np.concatenate(
-            (time_stamps_list, padding_time_stamp), axis=1
+        adjust_learning_rate(
+            optimizer, self.scheduler, epoch + 1, config, printout=False
         )
-        padding_mark = get_time_mark(whole_time_stamp, 1, self.config.freq)
-        return padding_mark
+        self.scheduler.step()
 
-    def validate(
-        self, valid_data_loader: DataLoader, series_dim: int, criterion: torch.nn.Module
-    ) -> float:
-        """
-        Validates the model performance on the provided validation dataset.
-        :param valid_data_loader: A PyTorch DataLoader for the validation dataset.
-        :param series_dim : The number of series data‘s dimensions.
-        :param criterion : The loss function to compute the loss between model predictions and ground truth.
-        :returns:The mean loss computed over the validation dataset.
-        """
-        config = self.config
-        total_loss = []
-        self.model.eval()
-        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
-        for input, target, input_mark, target_mark in valid_data_loader:
-            input, target, input_mark, target_mark = (
-                input.to(device),
-                target.to(device),
-                input_mark.to(device),
-                target_mark.to(device),
-            )
-            # decoder input
-            dec_input = torch.zeros_like(target[:, -config.horizon :, :]).float()
-            dec_input = (
-                torch.cat([target[:, : config.label_len, :], dec_input], dim=1)
-                .float()
-                .to(device)
-            )
+    def _init_model(self):
+        return PathformerModel(self.config)
 
-            output, balance_loss = self.model(input)
+    def _init_early_stopping(self):
+        return EarlyStopping(patience=self.config.patience)
 
-            target = target[:, -config.horizon :, :series_dim]
-            output = output[:, -config.horizon :, :series_dim]
-            loss = criterion(output, target).detach().cpu().numpy()
-            total_loss.append(loss)
+    def _process(self, input, target, input_mark, target_mark):
+        output, balance_loss = self.model(input)
 
-        total_loss = np.mean(total_loss)
-        self.model.train()
-        return total_loss
-
-    def forecast_fit(
-        self,
-        train_valid_data: pd.DataFrame,
-        *,
-        covariates: Optional[dict] = None,
-        train_ratio_in_tv: float = 1.0,
-        **kwargs,
-    ) -> "ModelBase":
-        """
-        Train the model.
-        :param train_valid_data: Time series data used for training and validation.
-        :param covariates: Additional external variables.
-        :param train_ratio_in_tv: Represents the splitting ratio of the training set validation set. If it is equal to 1, it means that the validation set is not partitioned.
-        :return: The fitted model object.
-        """
-        if covariates is None:
-            covariates = {}
-        series_dim = train_valid_data.shape[-1]
-        exog_data = covariates.get("exog", None)
-        if exog_data is not None:
-            train_valid_data = pd.concat([train_valid_data, exog_data], axis=1)
-
-        if train_valid_data.shape[1] == 1:
-            train_drop_last = False
-            self.single_forecasting_hyper_param_tune(train_valid_data)
-        else:
-            train_drop_last = True
-            self.multi_forecasting_hyper_param_tune(train_valid_data)
-
-        setattr(self.config, "task_name", "short_term_forecast")
-        self.model = PathformerModel(self.config)
-
-        print(
-            "----------------------------------------------------------",
-            self.model_name,
-        )
-        config = self.config
-        train_data, valid_data = train_val_split(
-            train_valid_data, train_ratio_in_tv, config.seq_len
-        )
-
-        self.scaler.fit(train_data.values)
-
-        if config.norm:
-            train_data = pd.DataFrame(
-                self.scaler.transform(train_data.values),
-                columns=train_data.columns,
-                index=train_data.index,
-            )
-
-        if train_ratio_in_tv != 1:
-            if config.norm:
-                valid_data = pd.DataFrame(
-                    self.scaler.transform(valid_data.values),
-                    columns=valid_data.columns,
-                    index=valid_data.index,
-                )
-            valid_dataset, valid_data_loader = forecasting_data_provider(
-                valid_data,
-                config,
-                timeenc=1,
-                batch_size=config.batch_size,
-                shuffle=True,
-                drop_last=False,
-            )
-
-        train_dataset, train_data_loader = forecasting_data_provider(
-            train_data,
-            config,
-            timeenc=1,
-            batch_size=config.batch_size,
-            shuffle=True,
-            drop_last=train_drop_last,
-        )
-
-        # Define the loss function and optimizer
-        if config.loss == "MSE":
-            criterion = nn.MSELoss()
-        elif config.loss == "MAE":
-            criterion = nn.L1Loss()
-        else:
-            criterion = nn.HuberLoss(delta=0.5)
-
-        optimizer = optim.Adam(self.model.parameters(), lr=config.learning_rate)
-
-        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-
-        self.early_stopping = EarlyStopping(patience=config.patience)
-        self.model.to(device)
-        total_params = sum(
-            p.numel() for p in self.model.parameters() if p.requires_grad
-        )
-
-        print(f"Total trainable parameters: {total_params}")
-        train_steps = len(train_data_loader)
-        scheduler = lr_scheduler.OneCycleLR(
-            optimizer=optimizer,
-            steps_per_epoch=train_steps,
-            pct_start=config.pct_start,
-            epochs=config.train_epochs,
-            max_lr=config.learning_rate,
-        )
-
-        for epoch in range(config.train_epochs):
-            self.model.train()
-            for i, (input, target, input_mark, target_mark) in enumerate(
-                train_data_loader
-            ):
-                optimizer.zero_grad()
-                input, target, input_mark, target_mark = (
-                    input.to(device),
-                    target.to(device),
-                    input_mark.to(device),
-                    target_mark.to(device),
-                )
-                # decoder input
-                dec_input = torch.zeros_like(target[:, -config.horizon :, :]).float()
-                dec_input = (
-                    torch.cat([target[:, : config.label_len, :], dec_input], dim=1)
-                    .float()
-                    .to(device)
-                )
-
-                output, balance_loss = self.model(input)
-
-                target = target[:, -config.horizon :, :series_dim]
-                output = output[:, -config.horizon :, :series_dim]
-                loss = criterion(output, target)
-
-                loss.backward()
-                optimizer.step()
-
-                if config.lradj == "TST":
-                    adjust_learning_rate(
-                        optimizer, scheduler, epoch + 1, config, printout=False
-                    )
-                    scheduler.step()
-
-            if train_ratio_in_tv != 1:
-                valid_loss = self.validate(valid_data_loader, series_dim, criterion)
-                self.early_stopping(valid_loss, self.model)
-                if self.early_stopping.early_stop:
-                    break
-            if config.lradj != "TST":
-                adjust_learning_rate(optimizer, scheduler, epoch + 1, config)
-
-    def forecast(
-        self,
-        horizon: int,
-        series: pd.DataFrame,
-        *,
-        covariates: Optional[dict] = None,
-    ) -> np.ndarray:
-        """
-        Make predictions.
-        :param horizon: The predicted length.
-        :param series: Time series data used for prediction.
-        :param covariates: Additional external variables
-        :return: An array of predicted results.
-        """
-        if covariates is None:
-            covariates = {}
-        series_dim = series.shape[-1]
-        exog_data = covariates.get("exog", None)
-        if exog_data is not None:
-            series = pd.concat([series, exog_data], axis=1)
-            if (
-                hasattr(self.config, "output_chunk_length")
-                and horizon != self.config.output_chunk_length
-            ):
-                raise ValueError(
-                    f"Error: 'exog' is enabled during training, but horizon ({horizon}) != output_chunk_length ({self.config.output_chunk_length}) during forecast."
-                )
-        if self.early_stopping.check_point is not None:
-            self.model.load_state_dict(self.early_stopping.check_point)
-
-        if self.config.norm:
-            series = pd.DataFrame(
-                self.scaler.transform(series.values),
-                columns=series.columns,
-                index=series.index,
-            )
-
-        if self.model is None:
-            raise ValueError("Model not trained. Call the fit() function first.")
-
-        config = self.config
-        series, test = split_time(series, len(series) - config.seq_len)
-        test = self.padding_data_for_forecast(test)
-
-        test_data_set, test_data_loader = forecasting_data_provider(
-            test, config, timeenc=1, batch_size=1, shuffle=False, drop_last=False
-        )
-
-        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-        self.model.to(device)
-        self.model.eval()
-
-        with torch.no_grad():
-            answer = None
-            while answer is None or answer.shape[0] < horizon:
-                for input, target, input_mark, target_mark in test_data_loader:
-                    input, target, input_mark, target_mark = (
-                        input.to(device),
-                        target.to(device),
-                        input_mark.to(device),
-                        target_mark.to(device),
-                    )
-                    dec_input = torch.zeros_like(
-                        target[:, -config.horizon :, :]
-                    ).float()
-                    dec_input = (
-                        torch.cat([target[:, : config.label_len, :], dec_input], dim=1)
-                        .float()
-                        .to(device)
-                    )
-                    output, balance_loss = self.model(input)
-
-                column_num = output.shape[-1]
-                temp = output.cpu().numpy().reshape(-1, column_num)[-config.horizon :]
-
-                if answer is None:
-                    answer = temp
-                else:
-                    answer = np.concatenate([answer, temp], axis=0)
-
-                if answer.shape[0] >= horizon:
-                    if self.config.norm:
-                        answer[-horizon:] = self.scaler.inverse_transform(
-                            answer[-horizon:]
-                        )
-                    return answer[-horizon:, :series_dim]
-
-                output = output.cpu().numpy()[:, -config.horizon :]
-                for i in range(config.horizon):
-                    test.iloc[i + config.seq_len] = output[0, i, :]
-
-                test = test.iloc[config.horizon :]
-                test = self.padding_data_for_forecast(test)
-
-                test_data_set, test_data_loader = forecasting_data_provider(
-                    test,
-                    config,
-                    timeenc=1,
-                    batch_size=1,
-                    shuffle=False,
-                    drop_last=False,
-                )
-
-    def batch_forecast(
-        self, horizon: int, batch_maker: BatchMaker, **kwargs
-    ) -> np.ndarray:
-        """
-        Make predictions by batch.
-
-        :param horizon: The length of each prediction.
-        :param batch_maker: Make batch data used for prediction.
-        :return: An array of predicted results.
-        """
-        if self.early_stopping.check_point is not None:
-            self.model.load_state_dict(self.early_stopping.check_point)
-
-        if self.model is None:
-            raise ValueError("Model not trained. Call the fit() function first.")
-        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-        self.model.to(device)
-        self.model.eval()
-
-        input_data = batch_maker.make_batch(self.config.batch_size, self.config.seq_len)
-        input_np = input_data["input"]
-        series_dim = input_np.shape[-1]
-
-        if input_data["covariates"] is None:
-            covariates = {}
-        else:
-            covariates = input_data["covariates"]
-        exog_data = covariates.get("exog")
-        if exog_data is not None:
-            input_np = np.concatenate((input_np, exog_data), axis=2)
-            if (
-                hasattr(self.config, "output_chunk_length")
-                and horizon != self.config.output_chunk_length
-            ):
-                raise ValueError(
-                    f"Error: 'exog' is enabled during training, but horizon ({horizon}) != output_chunk_length ({self.config.output_chunk_length}) during forecast."
-                )
-        if self.config.norm:
-            origin_shape = input_np.shape
-            flattened_data = input_np.reshape((-1, input_np.shape[-1]))
-            input_np = self.scaler.transform(flattened_data).reshape(origin_shape)
-
-        input_index = input_data["time_stamps"]
-        padding_len = (
-            math.ceil(horizon / self.config.horizon) + 1
-        ) * self.config.horizon
-        all_mark = self._padding_time_stamp_mark(input_index, padding_len)
-
-        answers = self._perform_rolling_predictions(horizon, input_np, all_mark, device)
-
-        if self.config.norm:
-            flattened_data = answers.reshape((-1, answers.shape[-1]))
-            answers = self.scaler.inverse_transform(flattened_data).reshape(
-                answers.shape
-            )
-
-        return answers[..., :series_dim]
-
-    def _perform_rolling_predictions(
-        self,
-        horizon: int,
-        input_np: np.ndarray,
-        all_mark: np.ndarray,
-        device: torch.device,
-    ) -> list:
-        """
-        Perform rolling predictions using the given input data and marks.
-
-        :param horizon: Length of predictions to be made.
-        :param input_np: Numpy array of input data.
-        :param all_mark: Numpy array of all marks (time stamps mark).
-        :param device: Device to run the model on.
-        :return: List of predicted results for each prediction batch.
-        """
-        rolling_time = 0
-        input_np, target_np, input_mark_np, target_mark_np = self._get_rolling_data(
-            input_np, None, all_mark, rolling_time
-        )
-        with torch.no_grad():
-            answers = []
-            while not answers or sum(a.shape[1] for a in answers) < horizon:
-                input, dec_input, input_mark, target_mark = (
-                    torch.tensor(input_np, dtype=torch.float32).to(device),
-                    torch.tensor(target_np, dtype=torch.float32).to(device),
-                    torch.tensor(input_mark_np, dtype=torch.float32).to(device),
-                    torch.tensor(target_mark_np, dtype=torch.float32).to(device),
-                )
-                output, balance_loss = self.model(input)
-                column_num = output.shape[-1]
-                real_batch_size = output.shape[0]
-                answer = (
-                    output.cpu()
-                    .numpy()
-                    .reshape(real_batch_size, -1, column_num)[
-                        :, -self.config.horizon :, :
-                    ]
-                )
-                answers.append(answer)
-                if sum(a.shape[1] for a in answers) >= horizon:
-                    break
-                rolling_time += 1
-                output = output.cpu().numpy()[:, -self.config.horizon :, :]
-                (
-                    input_np,
-                    target_np,
-                    input_mark_np,
-                    target_mark_np,
-                ) = self._get_rolling_data(input_np, output, all_mark, rolling_time)
-
-        answers = np.concatenate(answers, axis=1)
-        return answers[:, -horizon:, :]
-
-    def _get_rolling_data(
-        self,
-        input_np: np.ndarray,
-        output: Optional[np.ndarray],
-        all_mark: np.ndarray,
-        rolling_time: int,
-    ) -> Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
-        """
-        Prepare rolling data based on the current rolling time.
-
-        :param input_np: Current input data.
-        :param output: Output from the model prediction.
-        :param all_mark: Numpy array of all marks (time stamps mark).
-        :param rolling_time: Current rolling time step.
-        :return: Updated input data, target data, input marks, and target marks for rolling prediction.
-        """
-        if rolling_time > 0:
-            input_np = np.concatenate((input_np, output), axis=1)
-            input_np = input_np[:, -self.config.seq_len :, :]
-        target_np = np.zeros(
-            (
-                input_np.shape[0],
-                self.config.label_len + self.config.horizon,
-                input_np.shape[2],
-            )
-        )
-        target_np[:, : self.config.label_len, :] = input_np[
-            :, -self.config.label_len :, :
-        ]
-        advance_len = rolling_time * self.config.horizon
-        input_mark_np = all_mark[:, advance_len : self.config.seq_len + advance_len, :]
-        start = self.config.seq_len - self.config.label_len + advance_len
-        end = self.config.seq_len + self.config.horizon + advance_len
-        target_mark_np = all_mark[
-            :,
-            start:end,
-            :,
-        ]
-        return input_np, target_np, input_mark_np, target_mark_np
+        return {"output": output}
